entity: - Fix issues from MatchDetailDTO - (Needed to fix) There an error with DataInitializer

+++ Tournoi/src/main/java/be/TFTIC/Tournoi/dal/repositories/AddressRepository.java
@@ -3,5 +3,5 @@ package be.TFTIC.Tournoi.dal.repositories;
 import be.TFTIC.Tournoi.dl.entities.Address;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AddressRepository extends JpaRepository<Address, Integer> {
+public interface AddressRepository extends JpaRepository<Address, Long> {
 }
